changed file struicture in local flexible with additional code changes
